Enabled modules QuoteController.php
