--- app/Http/Requests/ClientPortal/ProcessQuotesInBulkRequest.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace App\Http\Requests\ClientPortal;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class ProcessQuotesInBulkRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            'quotes' => ['array'],
-        ];
-    }
-}
